move 'NewClient' function to 'utils' directory
